% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_mouse_convert.R
\name{human_mouse_convert}
\alias{human_mouse_convert}
\title{inter-convert between human and mouse gene symbols based on user's need}
\usage{
human_mouse_convert(input_genes, input_species)
}
\arguments{
\item{input_genes:}{the input gene list, must be a character vector}

\item{input_species:}{the input species that you want to convert from, could be "human" or "mouse". Must be in quotation marks. eg. if you put input_species = "human", then output species will be mouse}
}
\value{
a vector of genes of the output species
}
\description{
This function allows you to convert a list of human genes to mouse gene symbols and vice versa
}
\examples{
human_mouse_convert()
}
